+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Full documentation for rocFFT is available at [rocfft.readthedocs.io](https://ro
 - Reduced the number of syncs in encapsulated passes kernels, opportunistically improving 1D/2D/3D sizes.
 - Minor optimization for C2R 3D 100, 200 cube sizes.
 - Optimized some C2C/R2C 3D 64, 81, 100, 128, 200, 256 rectangular sizes.
+- When factoring, test to see if remaining length is explicitly
+  supported.
+- Explicitly add radix-7 lengths 14, 21, and 224 to list of supported
+  lengths.
 
 ## [rocFFT 1.0.10 for ROCm 4.1.0]